+++ src/indexing/indexStorageEvents.ts
@@ -1,0 +1,26 @@
+'use strict'
+
+import * as vscode from 'vscode'
+
+export class IndexStorageEvents implements vscode.Disposable {
+    
+    constructor() {
+        
+    }
+
+    public dispose() {
+
+    }
+
+    public onBeginRebuild() {
+
+    }
+
+    public onRebuild(path: string) {
+
+    }
+
+    public onEndRebuild() {
+
+    }
+}