+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
-## Release Notes
+# Release Notes
 
-### 0.0.1
+## 0.1.0
+### Added
+- Fuzzy search support
+- Unit tests
 
-Initial release
+### Changed
+- Settings prefix changed from `folderindex` to `folder-indexing`
+- Commands prefix changed from `folderindex` to `folder-indexing`
+
+## 0.0.1
+### Added
+- Initial release